refactor(routes): déplace les routes des tâches dans un module dédié/ feat(tasks): ajoute un endpoint PATCH pour les mises à jour partielles/ refactor(tasks): utilise des UUIDs pour l'identifiant des tâches
